refactor: move image-share handling onto the multimedia controller
