+++ AnkiDroid/src/main/java/com/ichi2/anki/noteeditor/NoteEditorMultimediaController.kt
@@ -21,9 +21,12 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
 import androidx.activity.result.ActivityResultLauncher
+import androidx.core.os.BundleCompat
 import androidx.lifecycle.lifecycleScope
 import com.ichi2.anki.CollectionManager.withCol
 import com.ichi2.anki.NoteEditorFragment
+import com.ichi2.anki.R
+import com.ichi2.anki.common.annotations.NeedsTest
 import com.ichi2.anki.compat.CompatHelper.Companion.getSerializableCompat
 import com.ichi2.anki.exception.MediaSizeLimitExceededException
 import com.ichi2.anki.launchCatchingTask
@@ -33,10 +36,14 @@ import com.ichi2.anki.multimedia.MultimediaResult
 import com.ichi2.anki.multimediacard.fields.EFieldType
 import com.ichi2.anki.multimediacard.fields.IField
 import com.ichi2.anki.servicelayer.NoteService
+import com.ichi2.anki.snackbar.showSnackbar
+import com.ichi2.utils.ContentResolverUtil
+import com.ichi2.utils.openInputStreamSafe
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import timber.log.Timber
+import java.io.File
 
 /**
  * Owns the multimedia capture lifecycle for a [NoteEditorFragment].
@@ -79,6 +86,89 @@ internal class NoteEditorMultimediaController(
                     true
                 }
             }
+    }
+
+    /**
+     * Handles an image-share intent: extracts the source [Uri], persists URI permission,
+     * copies it into the app's cache, and launches the image picker seeded with the
+     * cached file.
+     */
+    @NeedsTest("Test when the user directly passes image to the edit note field")
+    suspend fun handleImageIntent(data: Intent) {
+        val imageUri =
+            if (data.action == Intent.ACTION_SEND) {
+                BundleCompat.getParcelable(fragment.requireArguments(), Intent.EXTRA_STREAM, Uri::class.java)
+            } else {
+                data.data
+            }
+
+        if (imageUri == null) {
+            Timber.d("NoteEditor:: Image Uri is null")
+            fragment.showSnackbar(R.string.something_wrong)
+            return
+        }
+
+        try {
+            fragment
+                .requireContext()
+                .contentResolver
+                .takePersistableUriPermission(imageUri, Intent.FLAG_GRANT_READ_URI_PERMISSION)
+            Timber.d("Persisted URI permission for $imageUri")
+        } catch (e: SecurityException) {
+            Timber.w(e, "Unable to persist URI permission")
+        }
+
+        val cachedImagePath = copyUriToInternalCache(imageUri)
+        if (cachedImagePath == null) {
+            Timber.w("Failed to cache image")
+            fragment.showSnackbar(R.string.something_wrong)
+            return
+        }
+        val cachedUri = Uri.fromFile(File(fragment.requireContext().cacheDir, cachedImagePath))
+        launchImagePaste(cachedUri)
+    }
+
+    /**
+     * Copies a given [Uri] to the app's internal cache directory.
+     *
+     * This is necessary because URIs provided by other apps (e.g., WhatsApp, gallery apps) via
+     * `Intent` are usually content URIs with temporary permissions that are only valid
+     * in the originating context (like an Activity). Once passed to other components (like Fragments),
+     * these permissions may be lost, resulting in a SecurityException.
+     *
+     * By caching the file in internal storage and referencing it via a file URI,
+     * we ensure persistent access to the image without relying on external content providers.
+     *
+     * @param uri The [Uri] pointing to the external image content.
+     * @return The name of the cached file, or `null` if the operation failed.
+     */
+    private fun copyUriToInternalCache(uri: Uri): String? {
+        return try {
+            val context = fragment.requireContext()
+            val inputStream = context.contentResolver.openInputStreamSafe(uri) ?: return null
+
+            val fileName = ContentResolverUtil.getFileName(context.contentResolver, uri)
+            val cacheDir = context.cacheDir
+            val destFile = File(cacheDir, fileName)
+
+            val canonicalCacheDir = cacheDir.canonicalFile
+            val canonicalDestFile = destFile.canonicalFile
+
+            if (!canonicalDestFile.path.startsWith(canonicalCacheDir.path)) {
+                Timber.w("Rejected path due to directory traversal risk: $fileName")
+                return null
+            }
+
+            destFile.outputStream().use { output ->
+                inputStream.copyTo(output)
+            }
+
+            Timber.d("copyUriToInternalCache() copied to ${destFile.absolutePath}")
+            destFile.name
+        } catch (e: Exception) {
+            Timber.w(e, "Failed to copy URI to internal cache")
+            null
+        }
     }
 
     /**
